add Android build files and patches
Some functionality is skipped on Android due to lacking libc
support, and some things are simply not compiled (most notably gettext
support.) If you try to use those in a program, you'll get link errors.

Credit to https://android.googlesource.com/platform/external/bluetooth/glib for
the base of the makefiles, but I didn't use their patches, and re-derived
most of the content.